+++ migrations/2016_05_19_173453_create_menu_table.php
@@ -12,28 +12,31 @@ class CreateMenuTable extends Migration
      */
     public function up()
     {
-        Schema::create('menus', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name')->unique();
-            $table->timestamps();
-        });
+        if ( !Schema::hasTable('menus') ){
+            Schema::create('menus', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name')->unique();
+                $table->timestamps();
+            });
+        }
 
-        Schema::create('menu_items', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('menu_id')->nullable();
-            $table->string('title');
-            $table->string('url');
-            $table->string('target')->default('_self');
-            $table->string('icon_class')->nullable();
-            $table->string('color')->nullable();
-            $table->integer('parent_id')->nullable();
-            $table->integer('order');
-            $table->timestamps();
-        });
-
-        Schema::table('menu_items', function (Blueprint $table) {
-            $table->foreign('menu_id')->references('id')->on('menus')->onDelete('cascade');
-        });
+        if ( !Schema::hasTable('menu_items') ){
+            Schema::create('menu_items', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('menu_id')->nullable();
+                $table->string('title');
+                $table->string('url');
+                $table->string('target')->default('_self');
+                $table->string('icon_class')->nullable();
+                $table->string('color')->nullable();
+                $table->integer('parent_id')->nullable();
+                $table->integer('order');
+                $table->timestamps();
+            });
+            Schema::table('menu_items', function (Blueprint $table) {
+                $table->foreign('menu_id')->references('id')->on('menus')->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -43,7 +46,11 @@ class CreateMenuTable extends Migration
      */
     public function down()
     {
-        Schema::drop('menu_items');
-        Schema::drop('menus');
+        if ( Schema::hasTable('menu_items') ){
+            Schema::drop('menu_items');
+        }
+        if ( Schema::hasTable('menus') ){
+            Schema::drop('menus');
+        }
     }
 }
