+++ migrations/2016_10_21_190000_create_roles_table.php
@@ -13,12 +13,14 @@ class CreateRolesTable extends Migration
      */
     public function up()
     {
-        Schema::create('roles', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('name')->unique();
-            $table->string('display_name');
-            $table->timestamps();
-        });
+        if ( !Schema::hasTable('roles') ){
+            Schema::create('roles', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('name')->unique();
+                $table->string('display_name');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +30,8 @@ class CreateRolesTable extends Migration
      */
     public function down()
     {
-        Schema::drop('roles');
+        if ( Schema::hasTable('roles') ){
+            Schema::drop('roles');
+        }
     }
 }
