+++ migrations/2017_11_26_015000_create_user_roles_table.php
@@ -13,19 +13,21 @@ class CreateUserRolesTable extends Migration
      */
     public function up()
     {
-        Schema::create('user_roles', function (Blueprint $table) {
-            $type = DB::connection()->getDoctrineColumn(DB::getTablePrefix().'users', 'id')->getType()->getName();
-            if ($type == 'bigint') {
-                $table->bigInteger('user_id')->unsigned()->index();
-            } else {
-                $table->integer('user_id')->unsigned()->index();
-            }
+        if (!Schema::hasTable('user_roles')){
+            Schema::create('user_roles', function (Blueprint $table) {
+                $type = DB::connection()->getDoctrineColumn(DB::getTablePrefix().'users', 'id')->getType()->getName();
+                if ($type == 'bigint') {
+                    $table->bigInteger('user_id')->unsigned()->index();
+                } else {
+                    $table->integer('user_id')->unsigned()->index();
+                }
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->bigInteger('role_id')->unsigned()->index();
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->primary(['user_id', 'role_id']);
-        });
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->bigInteger('role_id')->unsigned()->index();
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->primary(['user_id', 'role_id']);
+            });
+        }
     }
 
     /**
